Rename formsFieldsMapper to formFieldsMapper
