+++ app/javascript/forms/mappers/formFieldsMapper.jsx
